--- src/main/java/com/wojciechbarwinski/demo/epic_board_games_shop/EmailSenderService.java
@@ -1,26 +1,0 @@
-package com.wojciechbarwinski.demo.epic_board_games_shop;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.mail.SimpleMailMessage;
-import org.springframework.mail.javamail.JavaMailSender;
-import org.springframework.stereotype.Service;
-
-@Service
-@RequiredArgsConstructor
-public class EmailSenderService {
-
-    @Value("${spring.mail.username}")
-    private String shopMail;
-
-    private final JavaMailSender mailSender;
-
-    public void sendSimpleMessage(String to, String subject, String text) {
-        SimpleMailMessage message = new SimpleMailMessage();
-        message.setFrom(shopMail);
-        message.setTo(to);
-        message.setSubject(subject);
-        message.setText(text);
-        mailSender.send(message);
-    }
-}
