+++ src/main/java/com/wojciechbarwinski/demo/epic_board_games_shop/messageSenders/MessageSenderPort.java
@@ -1,0 +1,8 @@
+package com.wojciechbarwinski.demo.epic_board_games_shop.messageSenders;
+
+import com.wojciechbarwinski.demo.epic_board_games_shop.entities.Order;
+
+public interface MessageSenderPort {
+
+    void sendSimpleMessageAfterOrderWasConfirm(Order order);
+}
