+++ src/main/resources/application.properties
@@ -30,9 +30,10 @@ orders.cleanup.not-payed.timer-minutes=1440
 spring.mail.host=smtp.gmail.com
 spring.mail.port=587
 
+spring.shopLink.order=http://localhost:8080/order/
+
 # How can I hide this data (mail and/or password) in git/gitHub?
 spring.mail.username=epicboardgameshop@gmail.com
-spring.mail.password=qoeyccnlzdwjosrr
 spring.mail.properties.mail.smtp.auth=true
 spring.mail.properties.mail.smtp.starttls.enable=true
 spring.mail.properties.mail.smtp.ssl.trust=smtp.gmail.com
